Split test suites for generating random and loading existing keys
